+++ CHANGELOG.md
@@ -1,15 +1,23 @@
 # Changelog
 
+## 0.1.x
+
+- add -(f)ocus to mapcat to show all drawn elements.
+- added -(r)eset to mapcat.
+
 ## 0.1.3
+
 - remove SingleInstance package.
 - add "clear" command to grep parser to remove all elements from map.
 
 ## 0.1.2
+
 - Support for several map windows.
 
 ## 0.1.1
+
 - Improve scrolling on Mac thanks to @RcusStackwalker.
 
 ## 0.1.0
-- Basic functionality
 
+- Basic functionality